refactor: split regen system into module with query/sys separation
Converts regen.py to regen/ package:
- query.py: esper data access (get_regen_candidates)
- sys.py: process logic using query layer
- __init__.py: exports process for state.py
