+++ ninjamagic/regen/__init__.py
@@ -1,0 +1,3 @@
+from ninjamagic.regen.sys import process
+
+__all__ = ["process"]
